LayoutTests/media/picture-in-picture-api-pip-events-expected.txt

This tests that events are fired correctly when a video element enters and exits the Picture-in-Picture mode.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EVENT(enterpictureinpicture)
EVENT(leavepictureinpicture)
END OF TEST
